+++ internal/provider/monitors_data_source.go
@@ -129,7 +129,7 @@ func (d *MonitorsDataSource) Schema(_ context.Context, _ datasource.SchemaReques
 							Computed:            true,
 						},
 						"expected_status_code": schema.StringAttribute{
-							MarkdownDescription: "Expected HTTP status code or pattern (e.g., `200`, `2xx`).",
+							MarkdownDescription: "Expected HTTP status code or pattern (e.g., `200`, `2xx`, `1xx-3xx`).",
 							Computed:            true,
 						},
 						"follow_redirects": schema.BoolAttribute{